+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Beanstalkd lib bundle
 
+## Unreleased
+
+* **[FIXED]** `Zlikavac32\BeanstalkdLibBundle\Command\Runnable\BeanstalkdServerControllerRunnable` crash on empty input line
+
 ## 0.1.0 (2019-05-05)
 
 * **[NEW]** First tagged version